Split and simplify
