+++ lua/autorun/cfc_contact_forms_init.lua
@@ -1,0 +1,1 @@
+include( "cfc_contact_forms/init.lua" )
